refactor(onboarding): extract ChecklistGroup component and useChecklistGroups hook from ChecklistSection (CU-31)
610L → 489L ChecklistSection; group accordion with expansion/completion logic in dedicated files.
