+++ apps/desktop/src/features/onboarding/ChecklistSection.tsx
@@ -29,7 +29,7 @@
  * dismiss affordances required for non-event items (FR-017).
  */
 
-import { useMemo, useState } from 'react';
+import { useState } from 'react';
 import { useNavigate, useRouterState } from '@tanstack/react-router';
 import {
   Check,
@@ -67,6 +67,8 @@ import {
   pageLabel,
   pagePath,
 } from './onboarding-labels';
+import { useChecklistGroups } from './useChecklistGroups';
+import { ChecklistGroup } from './ChecklistGroup';
 
 // Re-export the label/path helpers so existing callers (FindSpotlight, tests)
 // that import these from ChecklistSection still resolve.
@@ -112,59 +114,27 @@ export function ChecklistSection({
   // clickable — the overlay still dims the rest of the app.
   const spotlightActive = useActiveFindItem() != null;
 
-  // Whole-section collapse (FR-012) is backend-persisted; per-group manual
-  // toggles are session-local overrides on top of the route-driven default.
-  const [groupOverrides, setGroupOverrides] = useState<
-    Partial<Record<OnboardingPage, boolean>>
-  >({});
   // Section-header remove menu (T029): menu → one-line confirm → hide forever.
   const [menuOpen, setMenuOpen] = useState(false);
   const [removeConfirming, setRemoveConfirming] = useState(false);
 
-  const itemsByPage = useMemo(() => {
-    const map = new Map<OnboardingPage, OnboardingItemDto[]>();
-    if (state) {
-      for (const item of state.items) {
-        const list = map.get(item.page) ?? [];
-        list.push(item);
-        map.set(item.page, list);
-      }
-    }
-    return map;
-  }, [state]);
+  const currentPage = pageForPath(pathname);
+  const {
+    itemsByPage,
+    isGroupComplete,
+    isGroupExpanded,
+    groupProgress,
+    toggleGroup,
+  } = useChecklistGroups(state, currentPage);
 
   if (!state) return null;
 
-  const currentPage = pageForPath(pathname);
   const sectionExpanded = !state.flags.sidebarCollapsed;
   const { done, total } = state.progress;
   const progressText = m.onboarding_section_progress({ done, total });
 
-  const groupProgress = (page: OnboardingPage) =>
-    state.progress.perPage.find((p) => p.page === page) ?? {
-      done: 0,
-      total: 0,
-    };
-
-  const isGroupComplete = (page: OnboardingPage): boolean => {
-    const items = itemsByPage.get(page) ?? [];
-    return isChecklistGroupSettled(items);
-  };
-
-  // FR-007 auto-expand the current page's group; FR-031 takes precedence — a
-  // complete group stays a one-line done header even on its own page. A manual
-  // toggle (`groupOverrides`) always wins and never re-enables auto-expand.
-  const isGroupExpanded = (page: OnboardingPage): boolean => {
-    const override = groupOverrides[page];
-    if (override !== undefined) return override;
-    return page === currentPage && !isGroupComplete(page);
-  };
-
   const toggleSection = () =>
     void setOnboardingSection({ sidebarCollapsed: sectionExpanded });
-
-  const toggleGroup = (page: OnboardingPage) =>
-    setGroupOverrides((prev) => ({ ...prev, [page]: !isGroupExpanded(page) }));
 
   // Explicit removal (FR-013): hide the section (and its ring) forever, and
   // drop any spotlight that was open over a now-hidden item (spec edge case).
@@ -309,120 +279,22 @@ export function ChecklistSection({
             const completingHere = items.some((i) =>
               choreo.completingIds.has(i.itemId),
             );
-            // A settling item stays in the OPEN list (animating in place) until
-            // its completing window ends, then drops to the completed area.
-            const open = items.filter(
-              (i) =>
-                i.state === 'unchecked' || choreo.completingIds.has(i.itemId),
-            );
-            const completed = completedChecklistItems(items).filter(
-              (i) => !choreo.completingIds.has(i.itemId),
-            );
-            const g = groupProgress(page);
-            const complete = isGroupComplete(page);
             // Keep the group open through the choreography before FR-031
             // collapses it to its one-line done header (AS-6).
             const expanded = isGroupExpanded(page) || completingHere;
             return (
-              <div
+              <ChecklistGroup
                 key={page}
-                className={clsx(
-                  'pv-onb-checklist__group',
-                  complete && 'pv-onb-checklist__group--complete',
-                )}
-                data-testid="onb-checklist-group"
-              >
-                <button
-                  type="button"
-                  className="pv-onb-checklist__group-header"
-                  data-testid="onb-checklist-group-header"
-                  aria-expanded={expanded}
-                  onClick={() => toggleGroup(page)}
-                >
-                  {expanded ? (
-                    <ChevronDown size={13} aria-hidden />
-                  ) : (
-                    <ChevronRight size={13} aria-hidden />
-                  )}
-                  <span className="pv-onb-checklist__group-label">
-                    {pageLabel(page)}
-                  </span>
-                  {complete && (
-                    <Check
-                      size={13}
-                      aria-hidden
-                      className="pv-onb-checklist__group-done"
-                      data-testid="onb-checklist-group-done"
-                    />
-                  )}
-                  <span className="pv-onb-checklist__group-count">
-                    {g.done}/{g.total}
-                  </span>
-                </button>
-
-                {expanded && (
-                  <ul className="pv-onb-checklist__items">
-                    {open.map((item) => (
-                      <ChecklistItemRow
-                        key={item.itemId}
-                        item={item}
-                        idPrefix={idPrefix}
-                        completing={choreo.completingIds.has(item.itemId)}
-                        onJump={(jumpPage) =>
-                          void navigate({ to: pagePath(jumpPage) })
-                        }
-                      />
-                    ))}
-                    {completed.length > 0 && (
-                      <li
-                        className="pv-onb-checklist__completed"
-                        data-testid="onb-checklist-completed"
-                      >
-                        <ul className="pv-onb-checklist__items">
-                          {completed.map((item) => {
-                            const doneLabelId = `${idPrefix}-done-${item.itemId.replaceAll('.', '_')}`;
-                            return (
-                              <li
-                                key={item.itemId}
-                                className="pv-onb-checklist__item pv-onb-checklist__item--done"
-                                data-item-id={item.itemId}
-                                data-state={item.state}
-                              >
-                                <button
-                                  type="button"
-                                  role="checkbox"
-                                  aria-checked
-                                  aria-labelledby={doneLabelId}
-                                  className="pv-onb-checklist__check pv-onb-checklist__check--done"
-                                  onClick={() =>
-                                    void setOnboardingItemState(
-                                      item.itemId,
-                                      'unchecked',
-                                    )
-                                  }
-                                >
-                                  <Check
-                                    size={14}
-                                    aria-hidden
-                                    className="pv-onb-checklist__check-icon"
-                                  />
-                                </button>
-                                <span
-                                  id={doneLabelId}
-                                  className="pv-onb-checklist__item-label"
-                                  data-testid="onb-checklist-item-label"
-                                >
-                                  {itemLabel(item.itemId)}
-                                </span>
-                              </li>
-                            );
-                          })}
-                        </ul>
-                      </li>
-                    )}
-                  </ul>
-                )}
-              </div>
+                page={page}
+                items={items}
+                completingIds={choreo.completingIds}
+                expanded={expanded}
+                complete={isGroupComplete(page)}
+                groupProgress={groupProgress(page)}
+                idPrefix={idPrefix}
+                onToggle={() => toggleGroup(page)}
+                onJump={(jumpPage) => void navigate({ to: pagePath(jumpPage) })}
+              />
             );
           })}
         </div>
